requirements.txt
langchain>=0.3.27
langchain-core>=0.3.74
langchain-community>=0.3.27
langchain-text-splitters>=0.3.9
langchain-openai>=0.3.29
langchain-huggingface>=0.3.1


langgraph>=0.6.4
langgraph-checkpoint>=2.1.1
langgraph-prebuilt>=0.6.4
langgraph-sdk>=0.2.0

faiss-cpu>=1.11.0
numpy>=2.0.2

beautifulsoup4>=4.13.4
bs4>=0.0.2

httpx>=0.28.1
httpx-sse>=0.4.1

python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

hazm>=0.10.0

nltk>=3.9.1
